Enhance EXO self-signed certificate handling based on certificate hash (#71)
This update introduces a mechanism to check the certificate hash on each server startup. On the first startup, the certificate is automatically imported and its hash is saved. On subsequent startups, the hash is checked for changes. If the hash differs, the certificate is updated. In the case of an SSL fetch error, two modes are available: strict and non-strict. In strict mode, if the fetch fails for the first time, the server startup is aborted. In non-strict mode, the server will proceed with the current certificate even if the fetch fails.

(cherry picked from commit 3f21c6161cf9cf5ee1576527342381be73be92cb)